Selvedin: Fix for bug #277798. Fixing the issue that was reported by Bruce in regards to some serial number links still going out to WLN when they should be internal. Used this change to refactor domain to Long's in order to avoid exceptions in the future with serial numbers that exceed Integer.MAX_VALUE.
git-tfs-id: [https://tfsnpt.int.thomson.com:443/tfs/Cobalt_Collection/]$/Cobalt eBook Builder/Development;C223179

+++ ebookGenerator/src/com/thomsonreuters/uscl/ereader/gather/metadata/domain/DocMetadata.java
@@ -91,7 +91,7 @@ public class DocMetadata implements Serializable {
 	@Column(name = "SERIAL_NUMBER")
 	@Basic(fetch = FetchType.EAGER)
 	@XmlElement
-	Integer serialNumber;
+	Long serialNumber;
 	/**
 	 */
 
@@ -198,13 +198,13 @@ public class DocMetadata implements Serializable {
 
 	/**
 	 */
-	public void setSerialNumber(Integer serialNumber) {
+	public void setSerialNumber(Long serialNumber) {
 		this.serialNumber = serialNumber;
 	}
 
 	/**
 	 */
-	public Integer getSerialNumber() {
+	public Long getSerialNumber() {
 		return this.serialNumber;
 	}
